+++ packages/hydrogen/src/components/AddToCartButton/AddToCartButton.client.tsx
@@ -1,5 +1,6 @@
 import React, {ReactNode, useEffect, useState} from 'react';
 import {useCart} from '../CartProvider';
+import {useProduct} from '../ProductProvider';
 import {Props} from '../types';
 
 export interface AddToCartButtonProps {
@@ -9,7 +10,7 @@ export interface AddToCartButtonProps {
     value: string;
   }[];
   /** The ID of the variant. */
-  variantId: string;
+  variantId?: string | null;
   /** The item quantity. */
   quantity?: number;
   /** Any ReactNode elements. */
@@ -29,7 +30,7 @@ export function AddToCartButton<TTag extends React.ElementType = 'button'>(
 ) {
   const [addingItem, setAddingItem] = useState<boolean>(false);
   const {
-    variantId,
+    variantId: explicitVariantId,
     quantity = 1,
     attributes,
     children,
@@ -38,6 +39,19 @@ export function AddToCartButton<TTag extends React.ElementType = 'button'>(
     ...passthroughProps
   } = props;
   const {status, id, cartCreate, linesAdd} = useCart();
+  const product = useProduct();
+  const variantId =
+    explicitVariantId ??
+    product?.selectedVariant?.id ??
+    product?.variants?.find((variant) => variant.availableForSale)?.id ??
+    product?.variants?.[0]?.id ??
+    '';
+  const disabled =
+    explicitVariantId === null ||
+    variantId === '' ||
+    product?.selectedVariant === null ||
+    addingItem ||
+    passthroughProps.disabled;
 
   useEffect(() => {
     if (addingItem && status === 'idle') {
@@ -49,7 +63,7 @@ export function AddToCartButton<TTag extends React.ElementType = 'button'>(
     <>
       <button
         {...passthroughProps}
-        disabled={addingItem || passthroughProps.disabled}
+        disabled={disabled}
         onClick={() => {
           setAddingItem(true);
           if (!id) {