+++ packages/hydrogen/src/components/index.ts
@@ -45,10 +45,8 @@ export {ProductTitle} from './ProductTitle';
 export {ProductPrice} from './ProductPrice';
 export {ProductMetafield} from './ProductMetafield';
 export {SelectedVariantPrice} from './SelectedVariantPrice';
-export {SelectedVariantAddToCartButton} from './SelectedVariantAddToCartButton';
 export {SelectedVariantBuyNowButton} from './SelectedVariantBuyNowButton';
 export {SelectedVariantShopPayButton} from './SelectedVariantShopPayButton';
-export {SelectedVariantImage} from './SelectedVariantImage';
 export {SelectedVariantUnitPrice} from './SelectedVariantUnitPrice';
 export type {
   BuyNowButtonProps,