+++ packages/hydrogen/CHANGELOG.md
@@ -120,7 +120,7 @@ function SomeComponent() {
 - feat [breaking change]: Merge `/src/entry-server.jsx` entry point into `App.server.jsx`
 - feat: add a default virtual entry-client in `/@shopify/hydrogen/entry-client` that can be used in `index.html`
 - fix: prevent client components from being cached during development
-- refactor: the following components had their prop name renamed. Refer to the documentation or [#627](https://github.com/Shopify/hydrogen/issues/627) for more details.
+- refactor [breaking change]: the following components had their prop name renamed. Refer to the documentation or [#627](https://github.com/Shopify/hydrogen/issues/627) for more details.
   - `<ExternalVideo />`: renamed video prop to data
   - `<Video />`: renamed video prop to data
   - `<Image>`: renamed image prop to data
@@ -133,7 +133,9 @@ function SomeComponent() {
   - `<CartProvider>`: renamed cart prop to data
 - feat [breaking change]: remove `handleEvent` in favor of `handleRequest`.
 - feat [breaking change]: remove `assetHandler` parameter in the new `handleRequest`.
-- deps: update `react-helmet-async` to 1.2.3 and remove our custome type
+- deps: update `react-helmet-async` to 1.2.3 and remove our custom types
+- dx [breaking change]: `<SelectedVariantAddToCartButton />` has been removed; the `<AddToCartButton />` will now use the selectedVariant by default.
+- dx [breaking change]: remove the `<SelectedVariantImage />` component in favour of using `<Image data={product.selectedVariant.image} />`
 
 ## 0.10.1 - 2022-01-26